Move CloudEvents related code to camel-cloudevents
