+++ camel-k-knative/consumer/deployment/pom.xml
@@ -33,6 +33,10 @@
             <artifactId>camel-k-knative-consumer</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.apache.camel.k</groupId>
+            <artifactId>camel-k-cloudevents-deployment</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
             <artifactId>camel-quarkus-core-deployment</artifactId>
         </dependency>
