+++ camel-k-knative/consumer/runtime/pom.xml
@@ -48,6 +48,10 @@
             <groupId>org.apache.camel.k</groupId>
             <artifactId>camel-knative-http</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.apache.camel.k</groupId>
+            <artifactId>camel-k-cloudevents</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
