+++ camel-k-knative/runtime/pom.xml
@@ -49,6 +49,10 @@
             <groupId>org.apache.camel.k</groupId>
             <artifactId>camel-k-knative-impl</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.apache.camel.k</groupId>
+            <artifactId>camel-k-cloudevents</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
