+++ components/camel-cloudevents/src/main/java/org/apache/camel/component/cloudevents/CloudEvents.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.knative.spi;
+package org.apache.camel.component.cloudevents;
 
 import java.util.Arrays;
 import java.util.Collection;
