+++ components/camel-knative/camel-knative/src/main/java/org/apache/camel/component/knative/KnativeConfiguration.java
@@ -20,7 +20,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.camel.RuntimeCamelException;
-import org.apache.camel.component.knative.spi.CloudEvents;
+import org.apache.camel.component.cloudevents.CloudEvents;
 import org.apache.camel.component.knative.spi.KnativeEnvironment;
 import org.apache.camel.spi.UriParam;
 import org.apache.camel.spi.UriParams;
