+++ support/camel-k-runtime-bom/pom.xml
@@ -112,6 +112,16 @@
 
             <dependency>
                 <groupId>org.apache.camel.k</groupId>
+                <artifactId>camel-k-cloudevents</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.camel.k</groupId>
+                <artifactId>camel-k-cloudevents-deployment</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.camel.k</groupId>
                 <artifactId>camel-knative-api</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -128,6 +138,11 @@
             <dependency>
                 <groupId>org.apache.camel.k</groupId>
                 <artifactId>camel-knative-test</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.camel.k</groupId>
+                <artifactId>camel-cloudevents</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
